# lyshra-open-app-designer-api/src/main/resources/static/api/openapi.yaml
openapi: 3.1.0
info:
  title: Lyshra OpenApp Workflow Designer API
  description: |
    REST API for the Lyshra OpenApp Workflow Designer. This API provides endpoints for:
    - Workflow definition management (CRUD operations)
    - Workflow version management (activation, deprecation, rollback)
    - Workflow execution and monitoring
    - Processor metadata retrieval for the UI
    - Schema validation
    - User authentication and management
  version: 1.0.0
  contact:
    name: Lyshra OpenApp Team
    email: support@lyshra.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.lyshra.com
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/workflows:
    get:
      tags:
        - Workflows
      summary: List all workflows
      description: |
        Retrieve all workflow definitions with optional filtering.
        Supports filtering by organization, lifecycle state, and search term.
      operationId: getAllWorkflows
      parameters:
        - name: organization
          in: query
          description: Filter by organization namespace
          schema:
            type: string
            example: "acme-corp"
        - name: state
          in: query
          description: Filter by lifecycle state
          schema:
            $ref: '#/components/schemas/WorkflowLifecycleState'
        - name: search
          in: query
          description: Search term for name or description
          schema:
            type: string
            example: "order processing"
      responses:
        '200':
          description: List of workflow definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowDefinition'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Workflows
      summary: Create workflow
      description: Create a new workflow definition
      operationId: createWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowDefinitionCreateRequest'
            example:
              name: "Order Processing Workflow"
              description: "Handles order validation and fulfillment"
              organization: "acme-corp"
              module: "orders"
              category: "e-commerce"
              tags:
                priority: "high"
                team: "backend"
      responses:
        '201':
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDefinition'
          headers:
            Location:
              description: URL of the created resource
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/executions:
    get:
      tags:
        - Executions
      summary: List executions
      description: Get all workflow executions with optional filtering
      operationId: getAllExecutions
      parameters:
        - name: workflowDefinitionId
          in: query
          description: Filter by workflow definition ID
          schema:
            type: string
            format: uuid
        - name: workflowVersionId
          in: query
          description: Filter by workflow version ID
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filter by execution status
          schema:
            $ref: '#/components/schemas/ExecutionStatus'
        - name: startTime
          in: query
          description: Filter by start time (from)
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          description: Filter by start time (to)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowExecution'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/processors:
    get:
      tags:
        - Processors
      summary: List all processors
      description: |
        Get all available processor metadata for the workflow designer.
        Used to populate the processor palette in the UI.
      operationId: getAllProcessors
      parameters:
        - name: category
          in: query
          description: Filter by processor category
          schema:
            type: string
            example: "DATA_TRANSFORMATION"
        - name: search
          in: query
          description: Search term for processor name or description
          schema:
            type: string
            example: "json"
      responses:
        '200':
          description: List of processor metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProcessorMetadata'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/schemas:
    get:
      tags:
        - Schemas
      summary: List available schemas
      description: Get list of all available JSON schemas
      operationId: listSchemas
      responses:
        '200':
          description: List of schema information
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    type:
                      type: string
                    fileName:
                      type: string
                    available:
                      type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
...
  responses:
    BadRequest:
      description: Bad request - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2025-01-25T10:30:00Z"
            status: 400
            code: "VALIDATION_ERROR"
            message: "Request validation failed"
            details:
              fieldErrors:
                - field: "name"
                  message: "Workflow name is required"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2025-01-25T10:30:00Z"
            status: 401
            code: "UNAUTHORIZED"
            message: "Authentication required"
...
    Conflict:
      description: Conflict - resource already exists or state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: "2025-01-25T10:30:00Z"
            status: 409
            code: "CONFLICT"
            message: "Workflow with this name already exists"
  schemas:
...
    WorkflowDefinition:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        description:
          type: string
        organization:
          type: string
        module:
          type: string
        lifecycleState:
          $ref: '#/components/schemas/WorkflowLifecycleState'
        activeVersionId:
          type: string
          format: uuid
        tags:
          type: object
          additionalProperties:
            type: string
        createdBy:
          type: string
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    WorkflowLifecycleState:
      type: string
      enum:
        - DRAFT
        - ACTIVE
        - DEPRECATED
        - ARCHIVED
      description: |
        Lifecycle states:
        - DRAFT: Initial state, can be edited
        - ACTIVE: Published and can be executed
        - DEPRECATED: Still executable but no new executions recommended
        - ARCHIVED: Read-only, cannot be executed
    WorkflowDefinitionCreateRequest:
      type: object
      required:
        - name
        - organization
        - module
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        organization:
          type: string
          minLength: 1
          maxLength: 100
        module:
          type: string
          minLength: 1
          maxLength: 100
        category:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
...
    WorkflowExecution:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        workflowDefinitionId:
          type: string
          format: uuid
        workflowVersionId:
          type: string
          format: uuid
        workflowName:
          type: string
        versionNumber:
          type: string
        status:
          $ref: '#/components/schemas/ExecutionStatus'
        correlationId:
          type: string
        triggeredBy:
          type: string
        inputData:
          type: object
          additionalProperties: true
        outputData:
          type: object
          additionalProperties: true
        currentStepId:
          type: string
          format: uuid
        currentStepName:
          type: string
        errorCode:
          type: string
        errorMessage:
          type: string
        stepLogs:
          type: array
          items:
            $ref: '#/components/schemas/StepExecutionLog'
        retryCount:
          type: integer
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    ExecutionStatus:
      type: string
      enum:
        - PENDING
        - SCHEDULED
        - RUNNING
        - PAUSED
        - WAITING
        - COMPLETED
        - FAILED
        - CANCELLED
        - ABORTED
        - TIMED_OUT
...
    StepExecutionLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        stepId:
          type: string
          format: uuid
        stepName:
          type: string
        status:
          $ref: '#/components/schemas/StepExecutionStatus'
        branch:
          type: string
        attemptNumber:
          type: integer
        inputData:
          type: object
        outputData:
          type: object
        errorCode:
          type: string
        errorMessage:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    StepExecutionStatus:
      type: string
      enum:
        - PENDING
        - RUNNING
        - COMPLETED
        - FAILED
        - SKIPPED
        - CANCELLED
        - TIMED_OUT
...
    ProcessorMetadata:
      type: object
      properties:
        identifier:
          type: string
          description: Unique processor identifier
        name:
          type: string
        description:
          type: string
        category:
          type: string
        icon:
          type: string
          description: Bootstrap Icons name
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
        organization:
          type: string
        module:
          type: string
        version:
          type: string
        inputFields:
          type: array
          items:
            $ref: '#/components/schemas/ProcessorInputField'
        inputSchema:
          type: object
          description: JSON Schema for input data
        outputSchema:
          type: object
          description: JSON Schema for output data
        deprecated:
          type: boolean
          default: false
        deprecationMessage:
          type: string
    ProcessorInputField:
      type: object
      properties:
        name:
          type: string
        label:
          type: string
        type:
          $ref: '#/components/schemas/ProcessorFieldType'
        description:
          type: string
        required:
          type: boolean
          default: false
        defaultValue:
          description: Default value (type varies)
        placeholder:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/FieldOption'
        validation:
          type: object
          properties:
            pattern:
              type: string
            min:
              type: number
            max:
              type: number
            minLength:
              type: integer
            maxLength:
              type: integer
    ProcessorFieldType:
      type: string
      enum:
        - STRING
        - TEXT
        - NUMBER
        - BOOLEAN
        - SELECT
        - MULTI_SELECT
        - DATE
        - DATETIME
        - JSON
        - CODE
        - EXPRESSION
        - PASSWORD
        - URL
        - EMAIL
        - KEY_VALUE_PAIRS
    FieldOption:
      type: object
      properties:
        value:
          description: Option value
        label:
          type: string
        description:
          type: string

    # ============================================
    # Validation Schemas
    # ============================================
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - code
        - message
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          description: HTTP status code
        code:
          type: string
          description: Application error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details